requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
pillow>=9.0.0
matplotlib>=3.5.0
tqdm>=4.62.0
wandb>=0.13.0
kaggle>=1.5.12
kagglehub>=0.1.0
scikit-learn>=1.0.0
scipy>=1.7.0
lpips>=0.1.4
pytorch-fid>=0.2.1
tensorboard>=2.9.0